* client: Add basic state machine for client websocket connection
